add the ability to customize timer in multiplayer

+++ src/cpp/options_system.cpp
@@ -2,16 +2,17 @@
 #include <hpp/tiny_ecs_registry.hpp>
 #include <hpp/common.hpp>
 
-int decreaseTimer(Entity entity, int amount) {
+// change option timer type
+int decreaseTimer(Entity entity, float amount) {
 
 	auto& timer_count = registry.timer.get(entity).timerC;
-	timer_count = (timer_count - amount >= 20) ? timer_count - amount : 20;
+	timer_count = (timer_count - amount >= 20.f) ? timer_count - amount : 20.f;
 	return timer_count;
 }
 
-int increaseTimer(Entity entity, int amount) {
+int increaseTimer(Entity entity, float amount) {
 	auto& timer_count = registry.timer.get(entity).timerC;
-	timer_count = (timer_count < 50) ? timer_count + amount : 50;
+	timer_count = (timer_count < 50.f) ? timer_count + amount : 50.f;
 	return timer_count;
 }
